[#1244] - l18n: localize rhodes strings

+++ platform/shared/common/RhoTime.h
@@ -22,7 +22,7 @@ public:
 
     unsigned long toULong()
     {
-        return m_nativeTime;
+        return static_cast<unsigned long>(m_nativeTime);
     }
 
 protected:
